move line and tick info into seperate files
