+++ projects/ngx-d3timeline/src/lib/axis/axis-view-model.ts
@@ -1,0 +1,7 @@
+import { TickInfo } from './tick-info';
+import { Line } from './line';
+
+export interface AxisViewModel {
+  tickInfos: TickInfo[];
+  axisLine: Line;
+}
