+++ projects/ngx-d3timeline/src/lib/axis/line.ts
@@ -1,0 +1,6 @@
+export interface Line {
+  x1: number;
+  x2: number;
+  y1: number;
+  y2: number;
+}
