+++ projects/ngx-d3timeline/src/lib/axis/resources-axis/resources-axis.component.ts
