+++ projects/ngx-d3timeline/src/lib/axis/tick-info.ts
@@ -1,0 +1,4 @@
+export interface TickInfo {
+  label: string;
+  transform: string;
+}
