+++ projects/ngx-d3timeline/src/lib/axis/time-axis/time-axis.component.spec.ts
@@ -1,9 +1,9 @@
 import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { TimeAxisComponent } from './time-axis.component';
-import { TimelineView } from '../view/timeline-view';
+import { TimelineView } from '../../view/timeline-view';
 import { TimeAxisViewModel } from './time-axis-view-model';
-import { mockData } from '../mock-data';
-import { Orientation } from '../orientation';
+import { mockData } from '../../mock-data';
+import { Orientation } from '../../orientation';
 
 describe('TimeAxisComponent', () => {
   let fixture: ComponentFixture<TimeAxisComponent>;
