+++ projects/ngx-d3timeline/src/lib/axis/time-axis/time-axis.component.ts
